enhance: cleanup code, add comments
